pyproject.toml
[tool.poetry]
name = "dftimewolf"
version = "0.1.0"
description = ""
authors = ["Thomas Chopitea <tomchop@gmail.com>"]
license = "Apache"

[tool.poetry.scripts]
dftimewolf = 'dftimewolf.cli.dftimewolf_recipes:Main'

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
filelock = "*"
google-cloud-storage = "*"
google-cloud-logging = "*"
google-cloud-bigquery = "*"
PyPDF2 = "^2.6.0"
six = "^1.16.0"
timesketch-import-client = "^20210602"
pandas = "*"
google-cloud-error-reporting = "*"
cmd2 = "^2.4.2"
psutil = "^5.9.1"
prettytable = "^3.3.0"
vt-py = "^0.14.0"
boto3 = "^1.24.31"
python-magic = "^0.4.27"
turbinia-api-lib = "^1.0.2"
turbinia-client = "^1.0.7"
grr-api-client = "^3.4.7"
libcloudforensics = {git = "https://github.com/google/cloud-forensics-utils.git"}
docker = "^7.0.0"
setuptools = "^70.0.0"  # needed by docker
google-cloud-aiplatform = "^1.74.0"
google-generativeai = "^0.8.3"
ratelimit = "^2.2.1"
backoff = "^2.2.1"

[tool.poetry.extras]
turbinia_legacy = ["turbinia"]

[tool.poetry.group.dev.dependencies]
mock = "^4.0.3"
pylint = "^3.1.1"
coverage = "^6.4.2"
mypy = "*"
types-protobuf = "^3.19.22"
types-filelock = "^3.2.7"
types-PyYAML = "^6.0.10"
mkdocs = "^1.3.0"
graphviz = "^0.20"
pytype = "^2023.09.19"
ruff = "^0.3.4"
absl-py = "^2.1.0"

[tool.poetry.group.spannertelemetry]
optional = true

[tool.poetry.group.spannertelemetry.dependencies]
google-cloud-spanner = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
indent-width = 2
line-length = 80
